# aspnetcore-server/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Room Reservation API",
    "description" : "A Simple IP Address API",
    "version" : "0.1.9"
  },
  "servers" : [ {
    "url" : "http://localhost:5000/api"
  } ],
  "paths" : {
...
    "/reservations/{reservationId}" : {
      "get" : {
        "tags" : [ "reservations" ],
        "summary" : "Returns reservation by ID",
        "parameters" : [ {
          "name" : "reservationId",
          "in" : "path",
          "description" : "ID of a reservation",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeetingRoomReservation"
                }
              }
            }
          },
          "404" : {
            "description" : "Reservation not found"
          }
        }
      },
      "put" : {
        "tags" : [ "reservations" ],
        "summary" : "Updates an existing reservation",
        "parameters" : [ {
          "name" : "reservationId",
          "in" : "path",
          "description" : "ID of a reservation",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MeetingRoomReservation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Updated"
          },
          "400" : {
            "description" : "Bad request"
          },
          "404" : {
            "description" : "Room not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "reservations" ],
        "summary" : "Deletes a reservation",
        "parameters" : [ {
          "name" : "reservationId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "404" : {
            "description" : "Reservation not found"
          }
        }
      }
    },
...
    "/workingSpots/{workingSpotId}" : {
      "get" : {
        "tags" : [ "workingSpot" ],
        "summary" : "Returns a WorkingSpot info by ID",
        "parameters" : [ {
          "name" : "workingSpotId",
          "in" : "path",
          "description" : "ID of a workingSpot",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WorkingSpot"
                }
              }
            }
          },
          "404" : {
            "description" : "WorkingSpot not found"
          }
        }
      },
      "put" : {
        "tags" : [ "workingSpot" ],
        "summary" : "Updates an existing workingSpot",
        "parameters" : [ {
          "name" : "workingSpotId",
          "in" : "path",
          "description" : "ID of a workingSpot",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WorkingSpot"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Updated"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "WorkingSpot not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "workingSpot" ],
        "summary" : "Deletes a workingSpot",
        "parameters" : [ {
          "name" : "workingSpotId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "404" : {
            "description" : "WorkingSpot not found"
          }
        }
      }
    },
    "/workingSpotReservations" : {
      "get" : {
        "tags" : [ "workingSpotReservations" ],
        "summary" : "Returns a list of workingSpotReservation",
        "responses" : {
          "200" : {
            "description" : "An array of workingSpotReservations",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/WorkingSpotReservation"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "404" : {
            "description" : "WorkingSpot Reservations not found"
          }
        }
      },
      "post" : {
        "tags" : [ "workingSpotReservations" ],
        "summary" : "Creates a new workingSpotReservation",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WorkingSpotReservation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WorkingSpotReservation"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict"
          }
        }
      }
    },
    "/workingSpotsreservations/{workingSpotReservationId}" : {
      "get" : {
        "tags" : [ "workingSpotReservations" ],
        "summary" : "Returns a WorkingSpotReservation info by ID",
        "parameters" : [ {
          "name" : "workingSpotReservationId",
          "in" : "path",
          "description" : "ID of a workingSpotReservation",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WorkingSpotReservation"
                }
              }
            }
          },
          "404" : {
            "description" : "WorkingSpotReservation not found"
          }
        }
      },
      "put" : {
        "tags" : [ "workingSpotReservations" ],
        "summary" : "Update an existing workingSpot",
        "parameters" : [ {
          "name" : "workingSpotReservationId",
          "in" : "path",
          "description" : "ID of a workingSpotReservation",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WorkingSpotReservation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Updated"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "WorkingSpotReservation not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "workingSpotReservation" ],
        "summary" : "Deletes a workingSpotReservation",
        "parameters" : [ {
          "name" : "workingSpotReservationId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "404" : {
            "description" : "WorkingSpotReservation not found"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "MeetingRoom" : {
        "type" : "object",
        "properties" : {
          "meetingRoomId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "meetingRoomName" : {
            "type" : "string"
          },
          "meetingRoomEmail" : {
            "type" : "string",
            "format" : "email"
          },
          "meetingRoomLocation" : {
            "$ref" : "#/components/schemas/Location"
          },
          "locationId" : {
            "type" : "integer"
          },
          "meetingRoomImage" : {
            "$ref" : "#/components/schemas/MeetingRoomImage"
          },
          "meetingRoomReservations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MeetingRoomReservation"
            }
          }
        },
        "example" : {
          "meetingRoomLocation" : {
            "meetingRooms" : [ null, null ],
            "city" : {
              "cityName" : "cityName",
              "cityId" : 1
            },
            "locationId" : 6,
            "workingSpots" : [ {
              "workingSpotId" : 5,
              "locationId" : 5,
              "location" : null,
              "workingSpotReservations" : [ {
                "workingSpotId" : 7,
                "workingSpot" : null,
                "workingSpotReservationId" : 2,
                "reservationDate" : "2000-01-23",
                "user" : {
                  "jobId" : 2,
                  "password" : "",
                  "reservations" : [ {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  }, {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  } ],
                  "surname" : "surname",
                  "name" : "name",
                  "job" : {
                    "jobId" : 3,
                    "jobTitle" : "jobTitle"
                  },
                  "userId" : 9,
                  "email" : ""
                },
                "userId" : 1
              }, {
                "workingSpotId" : 7,
                "workingSpot" : null,
                "workingSpotReservationId" : 2,
                "reservationDate" : "2000-01-23",
                "user" : {
                  "jobId" : 2,
                  "password" : "",
                  "reservations" : [ {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  }, {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  } ],
                  "surname" : "surname",
                  "name" : "name",
                  "job" : {
                    "jobId" : 3,
                    "jobTitle" : "jobTitle"
                  },
                  "userId" : 9,
                  "email" : ""
                },
                "userId" : 1
              } ]
            }, {
              "workingSpotId" : 5,
              "locationId" : 5,
              "location" : null,
              "workingSpotReservations" : [ {
                "workingSpotId" : 7,
                "workingSpot" : null,
                "workingSpotReservationId" : 2,
                "reservationDate" : "2000-01-23",
                "user" : {
                  "jobId" : 2,
                  "password" : "",
                  "reservations" : [ {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  }, {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  } ],
                  "surname" : "surname",
                  "name" : "name",
                  "job" : {
                    "jobId" : 3,
                    "jobTitle" : "jobTitle"
                  },
                  "userId" : 9,
                  "email" : ""
                },
                "userId" : 1
              }, {
                "workingSpotId" : 7,
                "workingSpot" : null,
                "workingSpotReservationId" : 2,
                "reservationDate" : "2000-01-23",
                "user" : {
                  "jobId" : 2,
                  "password" : "",
                  "reservations" : [ {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  }, {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  } ],
                  "surname" : "surname",
                  "name" : "name",
                  "job" : {
                    "jobId" : 3,
                    "jobTitle" : "jobTitle"
                  },
                  "userId" : 9,
                  "email" : ""
                },
                "userId" : 1
              } ]
            } ]
          },
          "meetingRoomId" : 0,
          "meetingRoomName" : "meetingRoomName",
          "locationId" : 1,
          "meetingRoomImage" : {
            "imageWidth" : 7.457744773683766,
            "image" : "",
            "meetingRoom" : null,
            "meetingRoomId" : 1,
            "name" : "name",
            "meetingRoomImageId" : 6
          },
          "meetingRoomReservations" : [ null, null ],
          "meetingRoomEmail" : ""
        }
      },
      "MeetingRoomReservation" : {
        "type" : "object",
        "properties" : {
          "meetingRoomReservationId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "meetingDate" : {
            "type" : "string",
            "format" : "date"
          },
          "startTime" : {
            "type" : "string",
            "format" : "time"
          },
          "endTime" : {
            "type" : "string",
            "format" : "time"
          },
          "meetingTopic" : {
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/User"
          },
          "userId" : {
            "type" : "integer"
          },
          "meetingRoom" : {
            "$ref" : "#/components/schemas/MeetingRoom"
          },
          "meetingRoomId" : {
            "type" : "integer"
          }
        },
        "example" : {
          "meetingRoom" : null,
          "meetingRoomReservationId" : 4,
          "meetingRoomId" : 1,
          "meetingDate" : "2000-01-23",
          "startTime" : "startTime",
          "endTime" : "endTime",
          "meetingTopic" : "meetingTopic",
          "user" : null,
          "userId" : 7
        }
      },
      "User" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "name" : {
            "type" : "string"
          },
          "surname" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string",
            "format" : "email"
          },
          "password" : {
            "type" : "string",
            "format" : "password"
          },
          "job" : {
            "$ref" : "#/components/schemas/Job"
          },
          "jobId" : {
            "type" : "integer"
          },
          "reservations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MeetingRoomReservation"
            }
          }
        },
        "example" : {
          "jobId" : 2,
          "password" : "",
          "reservations" : [ {
            "meetingRoom" : null,
            "meetingRoomReservationId" : 4,
            "meetingRoomId" : 1,
            "meetingDate" : "2000-01-23",
            "startTime" : "startTime",
            "endTime" : "endTime",
            "meetingTopic" : "meetingTopic",
            "user" : null,
            "userId" : 7
          }, {
            "meetingRoom" : null,
            "meetingRoomReservationId" : 4,
            "meetingRoomId" : 1,
            "meetingDate" : "2000-01-23",
            "startTime" : "startTime",
            "endTime" : "endTime",
            "meetingTopic" : "meetingTopic",
            "user" : null,
            "userId" : 7
          } ],
          "surname" : "surname",
          "name" : "name",
          "job" : {
            "jobId" : 3,
            "jobTitle" : "jobTitle"
          },
          "userId" : 9,
          "email" : ""
        }
      },
      "MeetingRoomImage" : {
        "type" : "object",
        "properties" : {
          "meetingRoomImageId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "name" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string",
            "format" : "byte"
          },
          "imageWidth" : {
            "type" : "number"
          },
          "meetingRoomId" : {
            "type" : "integer"
          },
          "meetingRoom" : {
            "$ref" : "#/components/schemas/MeetingRoom"
          }
        },
        "example" : {
          "imageWidth" : 7.457744773683766,
          "image" : "",
          "meetingRoom" : null,
          "meetingRoomId" : 1,
          "name" : "name",
          "meetingRoomImageId" : 6
        }
      },
...
      "Job" : {
        "type" : "object",
        "properties" : {
          "jobId" : {
            "type" : "integer"
          },
          "jobTitle" : {
            "type" : "string"
          }
        },
        "example" : {
          "jobId" : 3,
          "jobTitle" : "jobTitle"
        }
      },
...
      "WorkingSpot" : {
        "type" : "object",
        "properties" : {
          "workingSpotId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "locationId" : {
            "type" : "integer"
          },
          "location" : {
            "$ref" : "#/components/schemas/Location"
          },
          "workingSpotReservations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WorkingSpotReservation"
            }
          }
        },
        "example" : {
          "workingSpotId" : 5,
          "locationId" : 5,
          "location" : null,
          "workingSpotReservations" : [ {
            "workingSpotId" : 7,
            "workingSpot" : null,
            "workingSpotReservationId" : 2,
            "reservationDate" : "2000-01-23",
            "user" : {
              "jobId" : 2,
              "password" : "",
              "reservations" : [ {
                "meetingRoom" : null,
                "meetingRoomReservationId" : 4,
                "meetingRoomId" : 1,
                "meetingDate" : "2000-01-23",
                "startTime" : "startTime",
                "endTime" : "endTime",
                "meetingTopic" : "meetingTopic",
                "user" : null,
                "userId" : 7
              }, {
                "meetingRoom" : null,
                "meetingRoomReservationId" : 4,
                "meetingRoomId" : 1,
                "meetingDate" : "2000-01-23",
                "startTime" : "startTime",
                "endTime" : "endTime",
                "meetingTopic" : "meetingTopic",
                "user" : null,
                "userId" : 7
              } ],
              "surname" : "surname",
              "name" : "name",
              "job" : {
                "jobId" : 3,
                "jobTitle" : "jobTitle"
              },
              "userId" : 9,
              "email" : ""
            },
            "userId" : 1
          }, {
            "workingSpotId" : 7,
            "workingSpot" : null,
            "workingSpotReservationId" : 2,
            "reservationDate" : "2000-01-23",
            "user" : {
              "jobId" : 2,
              "password" : "",
              "reservations" : [ {
                "meetingRoom" : null,
                "meetingRoomReservationId" : 4,
                "meetingRoomId" : 1,
                "meetingDate" : "2000-01-23",
                "startTime" : "startTime",
                "endTime" : "endTime",
                "meetingTopic" : "meetingTopic",
                "user" : null,
                "userId" : 7
              }, {
                "meetingRoom" : null,
                "meetingRoomReservationId" : 4,
                "meetingRoomId" : 1,
                "meetingDate" : "2000-01-23",
                "startTime" : "startTime",
                "endTime" : "endTime",
                "meetingTopic" : "meetingTopic",
                "user" : null,
                "userId" : 7
              } ],
              "surname" : "surname",
              "name" : "name",
              "job" : {
                "jobId" : 3,
                "jobTitle" : "jobTitle"
              },
              "userId" : 9,
              "email" : ""
            },
            "userId" : 1
          } ]
        }
      },
      "WorkingSpotReservation" : {
        "type" : "object",
        "properties" : {
          "workingSpotReservationId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "reservationDate" : {
            "type" : "string",
            "format" : "date"
          },
          "workingSpotId" : {
            "type" : "integer"
          },
          "workingSpot" : {
            "$ref" : "#/components/schemas/WorkingSpot"
          },
          "user" : {
            "$ref" : "#/components/schemas/User"
          },
          "userId" : {
            "type" : "integer"
          }
        },
        "example" : {
          "workingSpotId" : 7,
          "workingSpot" : null,
          "workingSpotReservationId" : 2,
          "reservationDate" : "2000-01-23",
          "user" : {
            "jobId" : 2,
            "password" : "",
            "reservations" : [ {
              "meetingRoom" : null,
              "meetingRoomReservationId" : 4,
              "meetingRoomId" : 1,
              "meetingDate" : "2000-01-23",
              "startTime" : "startTime",
              "endTime" : "endTime",
              "meetingTopic" : "meetingTopic",
              "user" : null,
              "userId" : 7
            }, {
              "meetingRoom" : null,
              "meetingRoomReservationId" : 4,
              "meetingRoomId" : 1,
              "meetingDate" : "2000-01-23",
              "startTime" : "startTime",
              "endTime" : "endTime",
              "meetingTopic" : "meetingTopic",
              "user" : null,
              "userId" : 7
            } ],
            "surname" : "surname",
            "name" : "name",
            "job" : {
              "jobId" : 3,
              "jobTitle" : "jobTitle"
            },
            "userId" : 9,
            "email" : ""
          },
          "userId" : 1
        }
      }
    }
  }
}